Add etch_a_sketch project
Tried adding a gitignore which ignores md files and gitmodules file, so that
they would not be on the public page, but I think that isn't worth the cost of
making it harder to pull down and modify this package elsewhere.

Even though these files are public, they're also available on the github page
of the project so it's not like there's any secrets, and if someone wanted to
see my workflow it would be transparent from what they see on the homepage.