Fix compile script, action, and downgrade tempalte
The template relied on a feature of pandoc that's only in more
recent versions, namely `elseif`, which doesn't work on 2.5,
the version that I'm able to install easily on GitHub actions.

The change to the action installs pandoc and runs the same
script I use locally.

The script is cleaned up according to feedback from shellcheck,
and in the last version I removed checks that it's running in
the correct directory since the goal now is to automate it.