Generate header links with javascript
This was done after like an hour of searching for a way to
auto-generate self links in pandoc, which I just couldn't
seem to find. It appears they do this on the official
pandoc manual, but my guess is they do it with a Haskell
plugin which I don't want to deal with.

I found the following google groups posts about fixing the
same problem:

- https://groups.google.com/g/pandoc-discuss/c/7Lx6jIYZgJY
- https://groups.google.com/g/pandoc-discuss/c/F-X7FiRah8c?pli=1